pyproject.toml
[tool.poetry]
name = "mark-processor"
version = "0.1.0"
description = ""
authors = ["neildaniel221270 <112754523+neildaniel221270@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "main.py", from="src"}]

[tool.poetry.dependencies]
python = ">=3.8.13,<3.10.4"
pandas = ">=2.0.2,<3.0.0"
numpy = ">=1.24.3,<2.0.0"
matplotlib = { version = ">=3.9.0,<4.0.0", python = ">=3.9,<3.10.4" }
seaborn = ">=0.1,<0.13.3"
scikit-learn = ">=1.2.2,<2.0.0"
tensorflow = "==2.10"
opencv-python-headless = ">=4.7.0.72,<5.0.0.0"
patchify = ">=0.2.3,<0.3.0"
networkx = ">=2.7.1,<4.0.0"
scikit-image = ">=0.21.0,<0.22.0"
skan = ">=0.1,<0.11.1"
typer = ">=0.12.3,<0.13.0"
plotly = ">=5.14.1,<6.0.0"
mlflow = "^2.14.1"
requests = ">=2.31.0,<3.0.0"
tqdm = ">=4.65.0,<5.0.0"
keras = ">=2.8.0,<3.0.0"
tensorflow-io-gcs-filesystem = ">=0.18.0, <0.38.0"
torch = ">=1.0.0"
torchvision = ">=0.15.1,<0.16.0"
pre-commit = ">=3.2.0,<4.0.0"
geopandas = ">=0.13.0,<0.14.0"
azureml-core = ">=1.50.0,<1.57.0"
azure-identity = ">=1.12.0,<2.0.0"
azure-ai-ml = ">=1.4.0,<2.0.0"
poetry2conda = ">=0.3.0,<0.4.0"
azureml-mlflow = "==1.56.0"
opencensus-ext-azure = "==1.1.13"
azureml-inference-server-http = ">=0.1.5"
fastapi = ">=0.95.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-cov = "^5.0.0"

[tool.poetry2conda]
name = "mark-processor"

[tool.mypy]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
